moved all our files into folders and made the makefile automatically compile any files inside the folders
