+++ commands/splitCmd.h
@@ -1,6 +1,6 @@
 #ifndef SOLITAIRE_SI_H
 #define SOLITAIRE_SI_H
-#include "linkedList.h"
+#include "../utils/linkedList.h"
 
 Node *splitDeck(int split, Node **head);
